requirements.txt
openai
langchain
langchain-community
streamlit
python-dotenv
PyPDF2
pandas
pycryptodome
